Move definitions for template SLB_cin_text.bt to a single file
